test(buzz-agent): share the subprocess harness across permission suites
The permission work copy-pasted the subprocess Harness, capturing LLM,
and permission-answer helper into permission_boundary.rs and duplicated
approve_permission into fake_llm.rs, on top of the copy already in
regressions.rs. Extract the shared pieces into tests/common/mod.rs and
include it with mod common from each suite; regressions.rs and
permission_boundary.rs now build on the shared Harness, fake_llm.rs
reuses the shared approve_permission. No test case is dropped or
weakened and no production code changes.

Fold in the two broker malformed-frame proofs (a frame carrying both
result and error, and one with a non-string method) into the e2e
adversarial matrix, so the frame-structure gate that only src-level unit
tests covered is now also observable at the live MCP-log seam.
